fixed: cancel button needs to be unclickable when submitting the form & move relative files in project-sttings folder in the playground
