+++ packages/playground/src/pages/sandbox-settings-project-members-page.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Spacer, Text, ListActions, SearchBox, Button } from '@harnessio/canary'
 import { SandboxLayout, SkeletonList, NoData } from '..'
 import { mockMemberData } from '../data/mockMembersData'
-import { MembersList } from '../components/members-list'
+import { MembersList } from '../components/project-settings/members-list'
 import { PlaygroundListSettings } from '../settings/list-settings'
 import { PaginationComponent } from '../components/pagination'
 import { useNavigate } from 'react-router-dom'
